Move the zip.h file inclusion
In some cases path to the zip.h path may be configured for libebook,
but not configured in other modules and this caused a "file not found"
error.  Since zip.h is not needed to declare EBook_EPUB class, it's
better to move zip.h file inclusion from ebook_epub.h to ebook_epub.cpp.