Move the link keeper of the new tab
To open pages in a new tab, the tabs of the navigation bar use
the context menu from the MainWindow. This context menu calls
handlers in the MainWindow, not in the tabs of the navigation bar.
Thus, tabs in the navigation bar need to be additionally passed
a link to the page. Previously, the m_newTabLinkKeeper variable
and the setTabKeeper and getNewTabLink functions in the ViewWindow
class were used for this. This is illogical and creates unnecessary
connectivity.

Now the m_newTabLink variable and the setNewTabLink and
getNewTabLink functions in the MainWindow class are used for this.